pyproject.toml
[project]
name = "blusky"
version = "0.1.0"
description = "Playing around with bluesky sdk"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.55",
    "black>=24.10.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "ruff>=0.7.4",
]
